s390/mem_detect: query diag500(4) to support QEMU/KVM memory devices
In order to support memory devices under QEMU/KVM, we have to prepare
our kernel virtual address space accordingly for the memory region that
hosts these devices.

Let's query the end address of the device memory region (which is always
placed above the SCLP maximum storage increment), and use it to fixup the
maximum possible pfn.

While we'd love to use one of the kvm_hypercall* variants, we have to
handle early exceptions.

Signed-off-by: David Hildenbrand <david@redhat.com>