virtio-mem: support VIRTIO_MEM_F_UNPLUGGED_INACCESSIBLE
The spec currently states that while unplugged memory should not be
read, the device has to allow for reading unplugged memory inside the
usable region. The primary motivation for this default handling was that
it simplified bring-up of virtio-mem in Linux.

In the meantime, known interfaces that allow access to physical memory
have been sanitized to not allow access to unplugged device memory:
- /dev/kmem has been removed
- /dev/mem does no longer allow to map virtio-mem device memory
- /proc/kcore will not read unplugged device memory (PageOffline)

For kdump, makedumpfile was updated to exclude PageOffline pages, just
like HW poisoned pages or unplugged memory pages part of memory blocks
added by the Hyper-V balloon. The only way kdump could still read unplugged
device memory would be when either using an old makedumpfile version
(unlikely, as new kernels regularly require adaptions and PageOffline
 support has been around before virtio-mem was merged) or dumping memory
via other tools or directly by copying /proc/vmcore (unlikely and currently
unsupported with virtio-mem).

For example, QEMU will soon indicate VIRTIO_MEM_F_UNPLUGGED_INACCESSIBLE
when not using anonymous memory but a file-backed, shmem, or hugetlbfs
memory backend to back device memory. Further, under s390x we'll
always set VIRTIO_MEM_F_UNPLUGGED_INACCESSIBLE due to lack of the shared
zeropage.

The device will indicate VIRTIO_MEM_F_UNPLUGGED_INACCESSIBLE and fail
device initialization if the driver does not indicate support. The
result is that Linux won't be able to make use of any memory without
this change. With this change, Linux will at least be able to (un)plug
in Linux memory block granularity. Print an info so this handling can
be identified.

In the context of the driver, we support
VIRTIO_MEM_F_UNPLUGGED_INACCESSIBLE, so let's unconditionally accept it.
We cannot 100% control if someone is going to do something nasty in
the kdump kernel.

In the future, we might provide a minimal virtio-mem driver to be used
inside the kdump kernel to catch and reject access to unplugged device
memory via /proc/vmcore via the oldmem_pfn_is_ram() callback -- to
support also dumping the vmcore without makedumpfile.

Cc: "Michael S. Tsirkin" <mst@redhat.com>
Cc: Jason Wang <jasowang@redhat.com>
Cc: Marek Kedzierski <mkedzier@redhat.com>
Cc: Hui Zhu <teawater@gmail.com>
Cc: Sebastien Boeuf <sebastien.boeuf@intel.com>
Cc: Pankaj Gupta <pankaj.gupta.linux@gmail.com>
Cc: Wei Yang <richard.weiyang@linux.alibaba.com>
Signed-off-by: David Hildenbrand <david@redhat.com>