mm: add folio_large_nr_pages()
Sometimes we might already know that we are dealing with a large folio,
and would want to avoid another folio_test_large() check.

Also, having a single way to extract the order of a large folios is
cleaner. So let's factor that out into a separate function.

Make use of it in internal.h and mm.h, where applicable.

Signed-off-by: David Hildenbrand <david@redhat.com>