mm/rmap: minimize folio->_nr_pages_mapped updates when batching PTE (un)mapping
It's not immediately obvious, but we can move the folio->_nr_pages_mapped
update out of the loop and reduce the number of atomic ops without
affecting the stats.

The important point to realize is that only removing the last PMD
mapping will result in _nr_pages_mapped going below ENTIRELY_MAPPED, not
the individual atomic_inc_return_relaxed() calls. Concurrent races with
removal of PMD mappings should be handled as expected, just like when
we would have such races right now on a single mapcount update.

In a simple munmap() microbenchmark [1] on 1 GiB of memory backed by the
same PTE-mapped folio size (only mapped by a single process such that they
will get completely unmapped), this change results in a speedup per folio
size on a x86-64 Intel machine of roughly:

* 16 KiB: 10%
* 32 KiB: 15%
* 64 KiB: 17%
* 128 KiB: 21%
* 256 KiB: 22%
* 512 KiB: 22%
* 1024 KiB: 23%
* 2048 KiB: 27%

[1] https://gitlab.com/davidhildenbrand/scratchspace/-/blob/main/pte-mapped-folio-benchmarks.c

Signed-off-by: David Hildenbrand <david@redhat.com>